# spec/contracts/modeling/rest-api.openapi.yaml
openapi: 3.1.0
info:
  title: Modeling API
  version: 1.0.0
  description: |
    Синхронный REST API для BC Modeling (docs/backend.md §5).

    Каждая операция ссылается на use case в `x-use-case` и на операцию таблицы
    доступа в `x-operation` (spec/nfr/modeling.yaml#security.authorization): роль
    проверяется на границе use case'а, а не матчером URL. Наполнено на этапе 2
    (2026-10-08) по одному endpoint'у на экспонированный сценарий; до этого
    `paths` был пуст намеренно — endpoint без `UC-*` нарушал бы трассируемость.

    Ошибки — `application/problem+json` (§8.1) с полем `code`: код инварианта
    (`INV-MDL-NNN`) или бизнес-ошибки (`RELATION_NOT_PERMITTED`, `MDL_*`).
    Класс отказа домена определяет статус: не найдено или скрыто списком доступа —
    `404`, конфликт состояния — `409`, нарушение правила — `422`, роль — `403`,
    очищенный снимок — `410`.

    Остаток XML (`raw_xml`, ADR-0017) наружу не отдаётся: это внутренний формат
    interchange.
servers:
  - url: /api/v1
    description: Относительный путь; приложение и фронтенд отдаются одним артефактом.
security:
  - bearerAuth: []
...
paths:
...
  /models/{id}/elements:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    post:
      tags: [content]
      operationId: createElement
      summary: Элемент в папке своего слоя; тип — фазы 1 (FR-07)
      x-use-case: UC-MDL-002
      x-operation: CreateElement
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [archiType, name]
              properties:
                archiType: { type: string, examples: ["archimate:ApplicationComponent"] }
                name:      { type: string, maxLength: 500 }
                folderId:  { type: string, format: uuid, description: "Без него — корень слоя." }
      responses:
        '201':
          description: Создан.
          content:
            application/json: { schema: { $ref: '#/components/schemas/Element' } }
        '409': { $ref: '#/components/responses/Conflict409' }
        '422': { $ref: '#/components/responses/UnprocessableEntity422' }
  /elements/{id}:
    parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
    patch:
      tags: [content]
      operationId: updateElement
      summary: Имя, документация, свойства; переименование видно на всех представлениях (FR-11)
      x-use-case: UC-MDL-002
      x-operation: UpdateElement
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ElementPatch' } }
      responses:
        '200':
          description: Изменён.
          content:
            application/json: { schema: { $ref: '#/components/schemas/Element' } }
    delete:
      tags: [content]
      operationId: deleteElement
      summary: Удалить; со связями — 409 с их перечнем в `relationships` (INV-MDL-004)
      x-use-case: UC-MDL-006
      x-operation: DeleteElement
      responses:
        '204': { description: Удалён вместе с размещениями на представлениях. }
        '409': { $ref: '#/components/responses/Conflict409' }
  /models/{id}/relationships:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    post:
      tags: [content]
      operationId: createRelationship
      summary: Связь по матрице ArchiMate 3.2; по запросу — и ребро на представлении
      x-use-case: UC-MDL-003
      x-operation: CreateRelationship
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [archiType, sourceId, targetId]
              properties:
                archiType: { type: string, examples: ["archimate:ServingRelationship"] }
                sourceId:  { type: string, format: uuid, description: "Элемент или связь." }
                targetId:  { type: string, format: uuid }
                name:      { type: string }
                view:
                  type: object
                  properties:
                    viewId:       { type: string, format: uuid }
                    sourceNodeId: { type: string, format: uuid }
                    targetNodeId: { type: string, format: uuid }
      responses:
        '201':
          description: Создана.
          content:
            application/json: { schema: { $ref: '#/components/schemas/Relationship' } }
        '200':
          description: Дубль того же типа между той же парой — существующая связь.
          content:
            application/json: { schema: { $ref: '#/components/schemas/Relationship' } }
        '422':
          description: |
            `RELATION_NOT_PERMITTED` с перечнем допустимых в `permitted` (INV-MDL-007)
            либо конец из другой модели (INV-MDL-004).
          content:
            application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
  /relationships/{id}:
    parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
    patch:
      tags: [content]
      operationId: updateRelationship
      summary: Имя, документация, свойства связи
      x-use-case: UC-MDL-003
      x-operation: UpdateRelationship
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ElementPatch' } }
      responses:
        '200':
          description: Изменена.
          content:
            application/json: { schema: { $ref: '#/components/schemas/Relationship' } }
    delete:
      tags: [content]
      operationId: deleteRelationship
      summary: Удалить вместе с рёбрами; примыкающая связь — 409 (INV-MDL-004)
      x-use-case: UC-MDL-003
      x-operation: DeleteRelationship
      responses:
        '204': { description: Удалена. }
        '409': { $ref: '#/components/responses/Conflict409' }
  /models/{id}/folders:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    post:
      tags: [content]
      operationId: createFolder
      summary: Пользовательская папка
      x-use-case: UC-MDL-006
      x-operation: ReorganizeTree
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parentId, name]
              properties:
                parentId: { type: string, format: uuid }
                name:     { type: string, maxLength: 500 }
      responses:
        '201':
          description: Создана в конце родителя (INV-MDL-005).
          content:
            application/json: { schema: { $ref: '#/components/schemas/Folder' } }
  /models/{id}/tree/{itemId}:
    parameters:
      - $ref: '#/components/parameters/ModelId'
      - { in: path, name: itemId, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [content]
      operationId: renameItem
      summary: Переименовать папку, элемент, связь или представление
      x-use-case: UC-MDL-006
      x-operation: ReorganizeTree
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [name], properties: { name: { type: string } } }
      responses:
        '204': { description: Переименовано. }
  /models/{id}/tree/move:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    post:
      tags: [content]
      operationId: moveToFolder
      summary: Перенос в папку; соседи не перенумеровываются, корни не меняются (INV-MDL-009)
      x-use-case: UC-MDL-006
      x-operation: ReorganizeTree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetFolderId, itemIds]
              properties:
                targetFolderId: { type: string, format: uuid }
                itemIds:        { type: array, items: { type: string, format: uuid } }
      responses:
        '204': { description: Перенесено целиком. }
        '422': { $ref: '#/components/responses/UnprocessableEntity422' }
  /models/{id}/tree/delete:
    parameters: [ { $ref: '#/components/parameters/ModelId' } ]
    post:
      tags: [content]
      operationId: deleteItems
      summary: Групповое удаление одной транзакцией
      x-use-case: UC-MDL-006
      x-operation: ReorganizeTree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemIds]
              properties:
                itemIds: { type: array, items: { type: string, format: uuid } }
      responses:
        '204': { description: Удалено целиком. }
        '409': { $ref: '#/components/responses/Conflict409' }
        '422': { $ref: '#/components/responses/UnprocessableEntity422' }
...
  /metamodel/relations:
    get:
      tags: [metamodel]
      operationId: relationTypes
      summary: Допустимые связи для упорядоченной пары типов; первая — по умолчанию
      x-use-case: UC-MDL-003
      parameters:
        - { in: query, name: source, required: true, schema: { type: string } }
        - { in: query, name: target, required: true, schema: { type: string } }
      responses:
        '200':
          description: Пусто — матрица не допускает ни одной.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/RelationType' } }

# Ещё не здесь: POST /views/{id}/auto-layout — этап 4 (docs/backend.md §5).
# Импорт и экспорт принадлежат BC interchange: contracts/interchange/rest-api.openapi.yaml.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT, выданный Keycloak. Роли VIEWER / ARCHITECT / ADMIN — в claim'ах токена (FR-28), группы — в `groups`.
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema: { type: string, maxLength: 64 }
      description: |
        Ключ дедупликации повторов. Тот же ключ с тем же телом возвращает результат
        первого выполнения; тот же ключ с другим телом — `409` (INV-MDL-003).
        Ключ принадлежит пользователю: чужой ключ с тем же значением не конфликтует.
    ModelId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  schemas:
    Problem:
      type: object
      description: |
        RFC 7807 Problem Details + поле `code` (docs/backend.md §8.1). Дополнительные поля
        по ситуации: `lockOwner` и `expiresAt` у чужой блокировки, `relationships` у удаления
        связанного элемента, `permitted` у недопустимой связи.
      required: [type, title, status, code]
      properties:
        type:     { type: string, format: uri }
        title:    { type: string }
        status:   { type: integer }
        detail:   { type: string }
        instance: { type: string }
        code:
          type: string
          description: Код инварианта (`INV-MDL-NNN`, domain/modeling/invariants.md) или иной бизнес-ошибки.
          examples: ["INV-MDL-004", "INV-MDL-006", "RELATION_NOT_PERMITTED"]
        lockOwner:     { type: string }
        relationships: { type: array, items: { type: string, format: uuid } }
        permitted:     { type: array, items: { type: string } }
...
    Property:
      type: object
      properties:
        key:   { type: string, maxLength: 200 }
        value: { type: string }
    Folder:
      type: object
      properties:
        id:         { type: string, format: uuid }
        parentId:   { type: string, format: uuid }
        archiId:    { type: string }
        name:       { type: string }
        folderType: { type: string, enum: [STRATEGY, BUSINESS, APPLICATION, TECHNOLOGY, MOTIVATION, IMPLEMENTATION_MIGRATION, OTHER, RELATIONS, DIAGRAMS] }
        sortOrder:  { type: integer, format: int64 }
    Element:
      type: object
      properties:
        id:            { type: string, format: uuid }
        folderId:      { type: string, format: uuid }
        archiId:       { type: string }
        archiType:     { type: string }
        layer:         { type: string }
        name:          { type: string }
        documentation: { type: string }
        properties:    { type: array, items: { $ref: '#/components/schemas/Property' } }
        supported:     { type: boolean, description: "false — opaque (FR-03): хранится, но не редактируется." }
        sortOrder:     { type: integer, format: int64 }
    ElementPatch:
      type: object
      description: Отсутствующее поле не меняется; `properties` — список целиком, пустой снимает все.
      properties:
        name:          { type: string }
        documentation: { type: string }
        properties:    { type: array, items: { $ref: '#/components/schemas/Property' } }
    Relationship:
      type: object
      properties:
        id:            { type: string, format: uuid }
        folderId:      { type: string, format: uuid }
        archiId:       { type: string }
        archiType:     { type: string }
        sourceId:      { type: string, format: uuid }
        targetId:      { type: string, format: uuid }
        name:          { type: string }
        documentation: { type: string }
        accessType:    { type: string, enum: [WRITE, READ, ACCESS, READ_WRITE] }
        directed:      { type: boolean }
        properties:    { type: array, items: { $ref: '#/components/schemas/Property' } }
        supported:     { type: boolean }
        sortOrder:     { type: integer, format: int64 }
        edgeId:        { type: string, format: uuid, description: "Ребро, если его просили создать." }
...
    RelationType:
      type: object
      properties:
        archiType: { type: string }
        type:      { type: string }
        byDefault: { type: boolean }
  responses:
...
    Conflict409:
      description: |
        Блокировки нет или она чужая, конкурентное сохранение (INV-MDL-006), повтор
        Idempotency-Key с другим телом (INV-MDL-003), удаление связанного элемента
        (INV-MDL-004), команда к удалённой модели (INV-MDL-002).
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
    UnprocessableEntity422:
      description: |
        Нарушение бизнес-правила: недопустимая связь (INV-MDL-007), тип не той фазы,
        папка чужой модели или не того корня (INV-MDL-009), запрещённый переход (INV-MDL-002).
      content:
        application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } }
...